making ir generation a pass.
